backend/requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==17.4.0
awscli==1.15.81
boto==2.49.0
botocore==1.10.80
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.2.2
defusedxml==0.5.0
Django==2.1.5
django-cors-headers==2.2.0
django-filter==1.1.0
django-mysql==2.4.1
django-ratelimit==1.1.0
django-storages==1.6.6
django-test-plus==1.0.22
djangorestframework==3.8.2
docutils==0.14
dry-rest-permissions==0.1.10
ecdsa==0.13
funcy==1.10.1
future==0.16.0
futures==3.1.1
gunicorn==19.8.1
idna==2.6
isodate==0.6.0
jmespath==0.9.3
langdetect==1.0.7
lxml==4.2.1
more-itertools==4.1.0
mysqlclient==1.3.14
oauthlib==2.1.0
Pillow==5.2.0
pkgconfig==1.3.1
pluggy==0.6.0
prettytable==0.7.2
py==1.5.3
pyasn1==0.4.4
pycparser==2.18
pycrypto==2.6.1
pycryptodomex==3.6.1
pyjwkest==1.4.0
PyJWT==1.6.4
pylibscrypt==1.7.1
pytest==3.5.0
pytest-django==3.1.2
python-dateutil==2.7.3
python-social-auth==0.3.6
python3-openid==3.1.0
python3-saml==1.4.1
pytz==2018.3
PyYAML==3.13
raven==6.9.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.13
scrypt==0.8.6
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
social-auth-steemconnect==0.0.3
steem==1.0.0
toolz==0.9.0
ujson==1.35
urllib3==1.22
voluptuous==0.11.1
w3lib==1.19.0
xmlsec==1.3.3
flex==6.13.2
drf-yasg
drf-yasg[flex]
drf-yasg[ssv]
swagger-spec-validator==2.4.1
django-silk==3.0.1
django-cors-headers==2.2.0
celery==4.3.0
pyfcm==1.4.5
redis==3.2.1
